configure: use same CPPFLAGS in kFreeBSD as Linux
046f081b46c8479820409cf8f530b988221bd15b reorganized the CPPFLAGS to no
longer add -D_POSIX_C_SOURCE unconditionally, but only on systems (e.g.,
glibc based ones) that require it.  As kFreeBSD uses glibc, it needs to
be treated similar.

Additionally, _BSD_SOURCE is turned on to enable some additional types
such as caddr_t, which are normally enabled on BSD but not with glibc.