Autodetect PowerPC vs. PowerPC64.
This is the same code as for x86_64.
This is necessary because uname returns PPC64 if the hardware
is 64 bit, however the userland can still be fully 32 bit.
In that case FFmpeg fails to compile because some macros in the
asm code are set up incorrectly.
For details see https://bugs.gentoo.org/show_bug.cgi?id=341235

Signed-off-by: Reimar Döffinger <Reimar.Doeffinger@gmx.de>